channel: Add ability to request sending the current state to the watcher
Mostly to demonstrate that this works, too and that we don't have to
wait for the acknowledgement of this message. Although you have to get
the state to the watcher within the challenge period, else the other
party could close the channel with an older state, since the watcher
doesn't know the most recent state.

Signed-off-by: Jens <jens@perun.network>