example: Emulate other participant,watcher and funder if in qemu
There is no easy way to implement a "standard" communication between
qemu and host applications (TCP), at least not that I'm aware of. In a
real project the application would know best how to get messages to
their destination. To demonstrate the full capability (minus TCP) the
example now emulates those three components and responds as if it had a
connection to them. It is implemented after the encoding (even though
that is harder/more annoying to do) and thus also tests the encoding
itself.

Signed-off-by: Jens <jens@perun.network>